requirements.txt
tensorflow==2.16.0
keras==3.3.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
MetaTrader5==5.0.45
google-cloud-aiplatform==1.38.0
google-cloud-storage==2.13.0
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.0
